+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 - New #232: Add `All` and `None` filters (@vjik)
 - Chg #233: Remove nullable types from `withFilter()` and `getFilter()` methods of `FilterableDataInterface` (@vjik)
 - Bug #234: Fix handling of `null` values in `IterableDataReader` (@vjik)
+- New #236: Add `PaginatorInterface::getFilter()` method (@vjik)
 
 ## 1.0.1 January 25, 2023